Rename package and use http utilities
