+++ apps/server-go-htmx/internal/api/api.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/spfave/projects-build/apps/server-go-htmx/pkg/web"
+	pHttp "github.com/spfave/projects-build/apps/server-go-htmx/pkg/http_utils"
 )
 
 type ApiServer struct {
@@ -20,17 +20,17 @@ func New(port int) *ApiServer {
 func (apiServer *ApiServer) Run() error {
 	server := &http.Server{
 		Addr:    fmt.Sprintf(":%d", apiServer.Port),
-		Handler: apiServer.RegisterHandlers(),
+		Handler: apiServer.RegisterRouteHandlers(),
 	}
 
 	return server.ListenAndServe()
 }
 
-func (apiServer *ApiServer) RegisterHandlers() http.Handler {
-	mux := http.NewServeMux()
-	mux.Handle("/api/v1/", http.StripPrefix("/api/v1", projectsRouter()))
-	mux.HandleFunc("/error", web.HandleError)
-	mux.HandleFunc("/", web.HandleNotFound)
+func (apiServer *ApiServer) RegisterRouteHandlers() http.Handler {
+	router := pHttp.NewRouter()
+	router.HandleSubroute("/api/v1", projectsRouter())
+	router.HandleFunc("/error", pHttp.HandlerError)
+	router.HandleFunc("/", pHttp.HandlerNotFound)
 
-	return mux
+	return router
 }
