+++ apps/server-go-htmx/pkg/http_utils/encoding.go
@@ -1,4 +1,4 @@
-package web
+package http_utils
 
 import (
 	"encoding/json"
